Site Version 3, Site Generator
Complete rewrite of the site so that the site is generated using a json file using generated.html as the template.

While the previous version of the site also used a json file, now instead of reading a hosted version of the json file and generating everything live, a  static site is generated with the file. This is a major improvement as it makes development easier. allows for faster loading, allows better customization per user, prevents downages if the json file cannot be read, and makes it whole system user friendly so others can actually use it and then customize their page without me there.

Go to generate.html, input the json and press "generate page". This will send you to a preview of the site, then press "download here" on the preview page to download.